// src/types/manager.type.ts
export interface Manager {
  userId: string; // Unique identifier for the building
  email?: string; // Name of the building
  fullName?: string; // Location of the building
  phoneNumber?: string; // URL of the building image
  dateOfBirth?: string; // Phone contact of the building
  buildingId?: string; // Phone contact of the building
}